+++ sync-for-commerce/CodatSyncCommerce/Models/Operations/CreateConnectionRequestBody.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncCommerce.Models.Operations
+namespace Codat.Sync.Commerce.Models.Operations
 {
     using Newtonsoft.Json;