+++ sync-for-commerce/CodatSyncCommerce/Utils/Utilities.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncCommerce.Utils
+namespace Codat.Sync.Commerce.Utils
 {
     using System;
     using System.Linq;
@@ -59,7 +59,7 @@ namespace CodatSyncCommerce.Utils
         {
             if (o == null)
                 return false;
-            return o.GetType().IsClass && (o.GetType().FullName ?? "").StartsWith("CodatSyncCommerce.Models");
+            return o.GetType().IsClass && (o.GetType().FullName ?? "").StartsWith("Codat.Sync.Commerce.Models");
         }
 
         // TODO: code review polyfilled for IsAssignableTo