+++ sync-for-commerce/RELEASES.md
@@ -93,3 +93,13 @@ Based on:
 - [csharp v2.3.0] sync-for-commerce
 ### Releases
 - [NuGet v2.3.0] https://www.nuget.org/packages/Codat.Sync.Commerce/2.3.0 - sync-for-commerce
+
+## 2023-11-13 01:25:21
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.0.0] sync-for-commerce
+### Releases
+- [NuGet v3.0.0] https://www.nuget.org/packages/Codat.Sync.Commerce/3.0.0 - sync-for-commerce